rename files based on project requirements
